src/optimizer/optimizer_template.py
"""
Optimizer Template
-----------------

This template provides a consistent structure for new optimizer classes for trading strategies.
Copy this file and fill in the specific logic for your optimizer.
"""

import os
import pandas as pd
from skopt.space import Real, Integer, Categorical
from src.optimizer.base_optimizer import BaseOptimizer

class OptimizerTemplate(BaseOptimizer):
    """
    Brief description of the optimizer and the strategy it optimizes.

    - Define the parameter search space.
    - Implement the optimization logic.
    - Handle result saving and plotting.
    """
    def __init__(self, data_dir='data/all', results_dir='results'):
        super().__init__(data_dir=data_dir, results_dir=results_dir)
        # Define the parameter search space
        self.search_space = [
            Integer(10, 30, name='example_period'),
            Real(1.0, 3.0, name='example_threshold'),
            # Add more parameters as needed
        ]
        self.default_params = {
            'example_period': 14,
            'example_threshold': 1.5,
        }

    def load_all_data(self):
        """
        Load all data files from the data directory.
        """
        # Example: self.data_files = [f for f in os.listdir(self.data_dir) if f.endswith('.csv')]
        super().load_all_data()

    def optimize_single_file(self, filename):
        """
        Run optimization for a single data file.
        Returns a dictionary with metrics and trades, or None if failed.
        """
        # Example implementation (replace with your own logic)
        try:
            df = pd.read_csv(os.path.join(self.data_dir, filename), index_col=0, parse_dates=True)
            # Run optimization logic here
            # result = self.run_optimization(df)
            result = {'metrics': {}, 'trades': []}  # Replace with actual results
            return result
        except Exception as e:
            print(f"Error optimizing {filename}: {e}")
            return None

    def plot_results(self, trades_df, metrics, save_path=None):
        """
        Plot the results of the optimization (trades, equity curve, etc.).
        """
        # Example: Use matplotlib to plot trades and metrics
        pass 