requirements.txt
pytest==7.4.3
python-dotenv==1.0.0
requests==2.31.0
numpy>=2.2.0
pandas>=2.2.0
scikit-optimize>=0.10.0
scikit-learn>=1.4.0  # Required by scikit-optimize
scipy>=1.12.0  # Required by scikit-optimize
TA-Lib>=0.4.0  # For technical indicators
python-dateutil>=2.8.0  # Required by pandas
pytz>=2024.1  # Required by pandas
python-binance==1.0.28
numpy>=1.24.0
ta>=0.10.0
python-dateutil>=2.8.2
requests>=2.31.0
aiohttp>=3.11.0
websockets>=15.0.0
pycryptodome>=3.22.0
backtrader>=1.9.78
sendgrid==6.12.0
flask>=3.0.0  # For web GUI
flask-login>=0.6.3  # For user authentication
yfinance>=0.2.61  # For Yahoo Finance data
lxml>=4.9.3  # For parsing HTML
beautifulsoup4>=4.12.2  # For parsing HTML
matplotlib>=3.7.0  # For plotting
seaborn>=0.12.2  # For advanced plotting
plotly>=5.15.0  # For interactive plotting
python-telegram-bot>=20.0  # For Telegram bot
cbpro>=4.0.0  # For Coinbase Pro API
ib_insync>=2.2.0  # For IBKR API
